% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{artemisColors}
\alias{artemisColors}
\title{Artemis Colors}
\usage{
artemisColors()
}
\value{
A \code{data.frame} with the following columns: \code{n}, \code{names}, \code{colors}, \code{r}, \code{g},
and \code{b}. The 3 first columns give the Artemis color number, its name, and
its equivalent in R. The final 3 columns give the r, g, and b values.
}
\description{
Returns a data frame with the standard artemis colors.
}
\examples{
artCol <- artemisColors()
plot(rep(1, nrow(artCol)), artCol$n, xlim = c(1, 2), type = "n")
text(rep(1, nrow(artCol)), artCol$n, labels = artCol$n, col = artCol$colors)
text(rep(1, nrow(artCol)), artCol$n, labels = artCol$names,
     col = artCol$colors, pos = 4, offset = 1)

}
\references{
Artemis website:
https://www.sanger.ac.uk/tool/artemis/
}
\author{
Lionel Guy
}
