% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_annotate.R
\name{auto_annotate}
\alias{auto_annotate}
\title{Auto-annotate dna_segs}
\usage{
auto_annotate(
  dna_seg_input,
  names = "gene",
  basic_mode = FALSE,
  locus_tag_pattern = NULL,
  keep_genes_only = TRUE,
  dna_seg = NULL,
  ...
)
}
\arguments{
\item{dna_seg_input}{Either a single \code{dna_seg} or a list of \code{dna_seg}
objects.}

\item{names}{A character string with the name of a \code{dna_seg} attribute, to
base the annotations on. If \code{dna_seg_input} is a single \code{dna_seg}, it
can also be a character vector with as many elements as there are rows in
that \code{dna_seg}.}

\item{basic_mode}{Logical. If \code{TRUE}, annotate the \code{dna_seg_input} using the
\code{names} argument, ignoring all other arguments and never generating spanning
annotations.}

\item{locus_tag_pattern}{A character string giving a pattern, used to
simplify names. To turn this off, use \code{locus_tag_pattern = ""}. See details.}

\item{keep_genes_only}{Logical. If \code{TRUE}, then for any row where \code{names} is
\code{"-"}, \code{"NA"}, or empty (\code{""}), no annotations will be made. See details.}

\item{dna_seg}{Deprecated, included for backwards compatibility.
When provided, replaces \code{dna_seg_input}.
it.}

\item{...}{Further arguments to be passed to the \code{annotation} function,
like \code{rot} or \code{col}.}
}
\value{
If \code{dna_seg_input} is a single \code{dna_seg}, then a single
\code{annotation} object will be returned.

If \code{dna_seg_input} is a list of \code{dna_seg} objects, a list of
\code{annotation} objects will be returned of equal length.
}
\description{
Annotate \code{dna_segs} in a smart way. This is especially designed for
\code{dna_segs} read from GenBank or EMBL files, but can be used for other
purposes as well. With the default arguments, it produces \code{annotation}
objects based on the \code{gene} attribute of \code{dna_segs}.
}
\details{
\code{keep_genes_only} is intended to be used with the \code{gene} column. When \code{TRUE},
it will only make annotations for \code{names} that are not 'empty' (\code{"-"},
\code{"NA"}, or \code{""}). When it is \code{FALSE} however, \code{locus_tag_pattern} becomes
relevant. For any element of \code{names} that is 'empty', it will take the \code{name}
attribute of the \code{dna_seg} and remove the \code{locus_tag_pattern} from it
(e.g. \code{Eco003456} becomes \code{003456}, with \code{Eco} as the \code{locus_tag_pattern}).
If \code{locus_tag_pattern} is left as \code{NULL} it will attempt to determine a
common prefix automatically. To turn this behavior off, use
\code{locus_tag_pattern = ""}.

If \code{names} refers to gene names, it will create spanning annotations for
operons or sequences of genes. For this to work, gene names have to be
consecutive and end with a number or capital letter.
}
\examples{
## Prepare dna_seg
names <- paste("Eco", sprintf("\%04d", 1:20), sep = "")
gene <- c("-", "atpC", "atpB", "atpA", "atp2", 
          "-", "-", "cda1", "cda2", "cda3",
          "vcx23", "vcx22", "vcx21", "cde20",
          "-", "gfrU", "gfrT", "gfrY", "gfrX", "gfrW")
ds <- dna_seg(data.frame(name = names,
                         start = (1:20) * 3,
                         end = (1:20) * 3 + 2,
                         strand = rep(1, 20),
                         gene = gene,
                         stringsAsFactors = FALSE))

## Original annotation
annot1 <- annotation(x1 = middle(ds), text = ds$gene, rot = 45)

## auto_annotate with various options
annot2 <- auto_annotate(ds)
annot3 <- auto_annotate(ds, keep_genes_only = FALSE,
                        locus_tag_pattern = "", rot = 75)
annot4 <- auto_annotate(ds, keep_genes_only = FALSE, col = "red")

## Plot
plot_gene_map(dna_segs = list(ds, ds, ds, ds),
              annotations = list(annot1, annot2, annot3, annot4))

}
\seealso{
\link{annotation}, \link{dna_seg}
}
\author{
Lionel Guy, Mike Puijk
}
