% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_data.R
\name{best_hit}
\alias{best_hit}
\title{Filter a comparison to include only the best hits for each query}
\usage{
best_hit(
  comparison,
  group_by = "name1",
  sort_order = c("bit_score", "aln_len", "per_id", "direction")
)
}
\arguments{
\item{comparison}{A \code{comparison} object to filter.}

\item{group_by}{A character string referring to a column in \code{comparison}
that holds the query names.}

\item{sort_order}{A character vector of column names that determine what the
best hit is, in order of importance. The first column is used to determine
the best hits per query, and each subsequent column is only used in case of
ties.}
}
\value{
A \code{comparison} object.
}
\description{
Takes a \code{comparison} object and filters it to include only the best hit
for each query. The query names are provided by \code{group_by}, and the best hit
is determined by sorting each set of queries based on a set of columns
provided by \code{sort_order}.
}
\details{
Designed to find the best hits per query from \code{comparison} objects
generated from tabular BLAST or DIAMOND results. For each column from
\code{sort_order}, the maximum (or minimum) value is taken for each unique query
name from \code{group_by}, in the order provided by \code{sort_order}, until each query
only 1 hit left. The maximum values are taken for each \code{sort_order} column,
unless the column provided is one of: \code{"mism"}, \code{"gaps"}, \code{"e_value"},
\code{"name1"}, \code{"name2"}, \code{"start1"}, \code{"start2"}, \code{"end1"}, or \code{"end2"}.
}
\examples{
## Read example blastp results
infile <- system.file('extdata/blastp_example1.tab', package = 'genoPlotR')

## comparison before filtering for best hits
blast_comparison <- read_comparison_from_blast(infile)
print_comparison(blast_comparison)

## Filter for best hits and print results
bh_comparison <- best_hit(blast_comparison)
print_comparison(bh_comparison)

}
\seealso{
\link{bidirectional_best_hit}, \link{comparison}, \link{run_blast}, \link{run_diamond}
}
\author{
Mike Puijk
}
