% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_data.R
\name{bidirectional_best_hit}
\alias{bidirectional_best_hit}
\title{Filter a comparison to include only bidirectional best hits}
\usage{
bidirectional_best_hit(
  comparison,
  other_direction,
  group_by1 = "name1",
  group_by2 = "name2"
)
}
\arguments{
\item{comparison}{A \code{comparison} object to filter.}

\item{other_direction}{A \code{comparison} object that the \code{comparison} to
filter is compared with.}

\item{group_by1}{A character string referring to a column in \code{comparison}
that holds its query names.}

\item{group_by2}{A character string referring to a column in
\code{other_direction} that holds its query names.}
}
\value{
A \code{comparison} object.
}
\description{
Takes a \code{comparison} object and filters it to include only
bidirectional best hits, with the use of a second \code{comparison},
provided by \code{other_direction}. Both \code{comparison} objects must be
filtered for best hits already (see \link{best_hit}).
}
\details{
The best hits from the first
\code{comparison} are only kept when their query-subject combinations can also
be found in the best hits from the second \code{comparison}. For example, take
a best hit in the first \code{comparison} with the query name \code{"geneA"} and
subject name \code{"geneB"}. In the \code{comparison} provided by
\code{other_direction}, the best hit for the query name \code{"geneB"} has to be the
subject name \code{"geneA"}. Only then is the hit kept as a bidirectional best
hit. The query names are provided by \code{group_by1} and \code{group_by2}, for the
\code{comparison} to filter and the other \code{comparison} respectively.
}
\examples{
## Read example blastp results
infile1 <- system.file('extdata/blastp_example1.tab', package = 'genoPlotR')

## comparison before filtering for best hits
blast_comparison1 <- read_comparison_from_blast(infile1)
print_comparison(blast_comparison1)

## Filter for best hits and print results
bh_comparison1 <- best_hit(blast_comparison1)
print_comparison(bh_comparison1)

## Repeat steps BLAST results in the other direction

infile2 <- system.file('extdata/blastp_example2.tab', package = 'genoPlotR')
blast_comparison2 <- read_comparison_from_blast(infile2)
bh_comparison2 <- best_hit(blast_comparison2)

## Filter for bidirectional best hits and print results
bbh_comparison1 <- bidirectional_best_hit(comparison = bh_comparison1,
                                         other_direction = bh_comparison2)
print_comparison(bbh_comparison1)

}
\seealso{
\link{best_hit}, \link{comparison}, \link{run_blast}, \link{run_diamond}
}
\author{
Mike Puijk
}
