% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPlotR.R
\docType{data}
\name{chrY_subseg}
\alias{chrY_subseg}
\title{Comparisons of subsegments of the Y chromosome in human and chimp}
\format{
\code{chrY_subseg}, a list of two data frame lists, representing the Y
segment in the two species and a comparison between them:
\describe{
\item{\code{dna_segs}}{A list of 2 \code{dna_seg} objects, containing 3 genes each.
Each exon and intron is a seperate feature (row) in the \code{dna_seg}.}
\item{\code{comparison}}{A list containing 1 \code{comparison} object.}
}
}
\usage{
data(chrY_subseg)
}
\description{
A subsegment of the Y chromosome in Homo sapiens and Pan troglodytes, to
illustrate support for exons and introns.
}
\examples{
data(chrY_subseg)
plot_gene_map(chrY_subseg$dna_segs, chrY_subseg$comparison,
              dna_seg_scale = TRUE, scale = FALSE)

}
\keyword{datasets}
