% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignments.R
\name{comparisons_from_dna_segs}
\alias{comparisons_from_dna_segs}
\title{Create comparisons between DNA segments}
\usage{
comparisons_from_dna_segs(
  dna_segs = NULL,
  seg_labels = NULL,
  files = NULL,
  mode = "full",
  tool = "blast",
  algorithm = "blastp",
  sensitivity = "default",
  output_path = NULL,
  all_vs_all = FALSE,
  filt_length = "auto",
  use_cache = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{dna_segs}{A list of \code{dna_seg} objects to create comparisons between.
Either \code{dna_segs} or \code{files} must be provided.}

\item{seg_labels}{A character vector containing DNA segment labels.}

\item{files}{A character vector, containing file paths to the FASTA or
GenBank files. The comparisons will be made between these files. Either
\code{dna_segs} or \code{files} must be provided.}

\item{mode}{Determines how the comparisons will be filtered.
\code{"besthit"}, \code{"bidirectional"}, or \code{"full"}. If mode is \code{"besthit"}, only the
best hit will be taken from each input query (see \link{best_hit}). If mode is
\code{"bidirectional"}, then hits are only kept if they are the best hits for
their query in both directions (see \link{bidirectional_best_hit}). \code{"full"} means
that all sequence alignment results are considered.}

\item{tool}{Choice of sequence alignment tool. Either \code{"blast"} or
\code{"diamond"}.}

\item{algorithm}{Choice of BLAST algorithm to run. One of: \code{"blastp"},
\code{"blastp-fast"}, \code{"blastp-short"}, \code{"tblastx"}, \code{"blastn"}, \code{"blastn-short"},
\code{"megablast"}, or \code{"dc-megablast"}.}

\item{sensitivity}{Choice of sensitivity option when running DIAMOND. One of:
\code{"fast"}, \code{"default"}, \code{"mid-sensitive"}, \code{"sensitive"}, \code{"more-sensitive"},
\code{"very-sensitive"}, or \code{"ultra-sensitive"}.}

\item{output_path}{Path to the folder that will contain the output files.
Both the sequence alignment result and the FASTA files used to make them will
be stored here.}

\item{all_vs_all}{Logical. If \code{TRUE}, sequence alignments will be performed
for every combination of the inputs, instead of just the ones necessary for
plotting. Note that this can take a long time, so use with caution.}

\item{filt_length}{A number indicating the minimum length required for hits,
or \code{"auto"}. If \code{"auto"}, it will be determined based on the choice of \code{tool}
and \code{algorithm} (150 for DIAMOND or any blastp algorithm, 450 for tblastx,
900 for any blastn algorithm).}

\item{use_cache}{Logical. If \code{FALSE}, it will never check for existing files.
This includes the FASTA files used as input for sequcence alignment, the
database files used by DIAMOND and BLAST, and the sequence alignment results
themselves.}

\item{verbose}{Logical. If \code{TRUE}, reports timings when creating new files.}

\item{...}{Arguments to pass to other functions (the functions executing
the sequence alignments tools, \link{run_blast}, and \link{run_diamond}, and the
function for parsing BLAST/ DIAMOND results,
\link{read_comparison_from_blast}).}
}
\value{
A list of \code{comparison} objects.
}
\description{
Create a list of \code{comparison} objects from a list of \code{dna_seg} objects or
files by parsing (and executing) sequence alignments. If these files already
exist, then those will be parsed. If not, DIAMOND or a BLAST program can be
executed to generate the sequence alignment results between the \code{dna_seg},
with respect to the order of \code{dna_segs}. Executing DIAMOND or BLAST requires
that the command-line implementations of these tools are installed.
}
\details{
Unless \code{use_cache} is set to \code{FALSE}, this function will look for the files
required using a combination of the \code{seg_labels} (if these are provided), and
the names of the \code{dna_segs} or \code{files} that were provided as input. If it
cannot find sequence alignment results in the form of \code{"query_subject"}
(or to put it differently, \code{"dna_seg1_dna_seg2"}), then it will run DIAMOND
or BLAST to generate these results. Using this system, it also looks for the
FASTA files required as input for the sequence alignment.

If \code{output_path} is left as \code{NULL}, the current working directory will be
used instead.
}
\examples{
\dontrun{
## Comparisons from a vector of GenBank files using DIAMOND
comparisons <- comparisons_from_dna_segs(
  files = c("genome1.gb", "genome2.gb", "genome2.gb"),
  tool = "diamond",
  output_path = "output/diamond",
  sensitivity = "very-sensitive",
  verbose = TRUE
)
}

}
\seealso{
\link{run_blast}, \link{run_diamond}, \link{dna_seg_to_fasta}, \link{genbank_to_fasta}
}
\author{
Mike Puijk
}
