% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_seg.R
\name{dna_seg}
\alias{dna_seg}
\alias{as.dna_seg}
\alias{is.dna_seg}
\title{dna_seg class and methods}
\usage{
dna_seg(x, ...)

as.dna_seg(
  df,
  col = "grey20",
  fill = "grey80",
  lty = 1,
  lwd = 1,
  pch = 8,
  cex = 1,
  gene_type = "arrows",
  region_plot = "NA",
  ordered = TRUE
)

is.dna_seg(dna_seg)
}
\arguments{
\item{x}{An object to be coerced. Can be a \code{data.frame}, \code{data.table},
or \code{list} object. See details for necessary columns.}

\item{...}{Arguments to pass to \link{dna_seg}, see arguments below.}

\item{df}{Same as \code{x}.}

\item{col}{A character vector of colors, of either length one or the same
length as \code{x}. Determines the outline of features, or their overall
color for shapes (see \link{gene_types}) that have no fill color (e.g. text,
points).}

\item{fill}{Same as \code{col}, but determines the fill color of features.}

\item{lty}{A vector of either length 1 or the same length as \code{x}. Determines
the line type (see \link{par}) of the features.}

\item{lwd}{A numeric vector of either length 1 or the same length as \code{x}.
Determines the line width (see \link{par}) of the features.}

\item{pch}{A vector of either length 1 or the same length as \code{x}. Determines
the shape (see \link{points}) of the features when they are represented by points.}

\item{cex}{A numeric vector of either length 1 or same the length as \code{x}.
Determines the size multiplier of features when they are represented by
text or points.}

\item{gene_type}{A character vector of either length 1 or same the length as
\code{x}. Determines the gene type (i.e. shape) of the features. See \link{gene_types}.}

\item{region_plot}{A character vector of either length 1 or the same length
as \code{x}. Determines if features are going to be plotted when going for
regional plotting. See \link{plot_gene_map} for more details.}

\item{ordered}{Logical. If \code{TRUE}, orders the \code{dna_seg} on the \code{start} and
\code{end} columns.}

\item{dna_seg}{An object to test.}
}
\value{
\code{as.dna_seg} and \code{dna_seg} both return a \code{dna_seg} object.

\code{is.dna_seg} returns a logical.
}
\description{
\code{dna_seg} objects are collections of genes or other elements along a genome,
to be represented on a map. These functions are class functions to create,
convert, and test \code{dna_seg} objects.
}
\details{
A \code{dna_seg} object is a \code{data.table} that must start with the columns \code{name},
\code{start}, \code{end}, and \code{strand}. The arguments listed above add more
columns, but only if those were not already present in the data provided by
\code{x}. Any number of additional columns can be added to \code{dna_seg} objects, to
act as metadata or as ways to group or identify features.

\code{dna_seg} and \code{as.dna_seg} can both be used to create \code{dna_seg} objects from
list, \code{data.frame}, or \code{data.table} objects. Alternatively,
\link{read_dna_seg_from_file} can be used to create them from a file.

These objects inherit from \code{data.table}, and
can therefore be manipulated using \code{data.table} syntax.

\code{is.dna_seg} returns \code{TRUE} if the object tested is a \code{dna_seg} object.
}
\examples{
## Generate data
names1 <- c("feat1", "feat2", "feat3")
starts1 <- c(2, 1000, 1050)
ends1 <- c(600, 800, 1345)
strands1 <- c("-", -1, 1)
cols1 <- c("blue", "grey", "red")

## Create data.frame
df1 <- data.frame(name = names1, start = starts1, end = ends1,
                  strand = strands1, col = cols1)

## Create dna_seg
dna_seg1 <- dna_seg(df1)
dna_seg1
as.dna_seg(df1)

## Test dna_seg
is.dna_seg(dna_seg1)

}
\seealso{
\link{read_dna_seg}, \link{trim.dna_seg}, \link{reverse.dna_seg}, \link{gene_types},
\link{update_dna_segs}
}
\author{
Lionel Guy, Mike Puijk
}
