% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignments.R
\name{dna_seg_to_fasta}
\alias{dna_seg_to_fasta}
\title{Convert dna_seg objects to FASTA files}
\usage{
dna_seg_to_fasta(
  dna_seg_input,
  output_path = NULL,
  id = "locus_id",
  file_names = NULL,
  unique_ids = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{dna_seg_input}{Either a single \code{dna_seg} object or a list of \code{dna_seg}
objects,}

\item{output_path}{Path to the folder that will contain the output files.}

\item{id}{The \code{dna_seg} column to use as the header for the FASTA sequences.}

\item{file_names}{A character vector of file names to use. If provided,
must be the same length as the amount of provided \code{dna_seg} objects.}

\item{unique_ids}{Logical. If \code{TRUE}, all values in the column provided by
\code{id} have to be unique. If \code{FALSE} it will use only the first sequence for
each unique value of \code{id}.}

\item{verbose}{Logical. If \code{TRUE}, report the time when it starts writing
to the output file, as well as a potential warning when the \code{id} column has
duplicate values.}

\item{...}{Arguments to pass to \link{fwrite}.}
}
\value{
Returns nothing (invisible \code{NULL}).
}
\description{
Takes the (non-boundary) features from \code{dna_seg_input} with sequences
and makes a FASTA file from them.
}
\details{
If \code{output_path} is left as \code{NULL}, the current working directory will be
used instead.
}
\examples{
\dontrun{
dna_seg_to_fasta(dna_seg_input = list(dna_seg1, dna_seg2), 
                 output_path = "path/to/output/folder")
}

}
\seealso{
\link{genbank_to_fasta}, \link{run_blast}, \link{run_diamond}
}
\author{
Mike Puijk
}
