% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignments.R
\name{genbank_to_fasta}
\alias{genbank_to_fasta}
\title{Convert GenBank files to FASTA files}
\usage{
genbank_to_fasta(
  file,
  id_tag = "locus_tag",
  take_origin = FALSE,
  out = NULL,
  verbose = FALSE
)
}
\arguments{
\item{file}{A GenBank file to convert.}

\item{id_tag}{A character string, denotes which CDS field to use as the
header of the FASTA file sequences. This field must exist in all CDS tags
present in the GenBank file.}

\item{take_origin}{Logical. If \code{TRUE}, will take the ORIGIN field(s) to
create a FASTA file containing nucleotide sequences instead.}

\item{out}{Output file path.}

\item{verbose}{Logical. If \code{TRUE}, report the time when it starts writing
to the output file.}
}
\value{
Returns nothing (invisible \code{NULL}).
}
\description{
Takes the \code{CDS} tags with \code{translation} fields and creates a FASTA file from
them, or takes the ORIGIN field(s) to create a FASTA file containing
nucleotide sequences instead.
}
\details{
If \code{out} is left as \code{NULL}, the current working directory will be used
instead.
}
\examples{
\dontrun{
genbank_to_fasta(file = "genome1.gb", out = "genome1")
}

}
\seealso{
\link{dna_seg_to_fasta}, \link{run_blast}, \link{run_diamond}
}
\author{
Mike Puijk
}
