% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_types.R
\name{gene_types}
\alias{gene_types}
\title{Lists gene_types for dna_seg objects}
\usage{
gene_types(auto = TRUE)
}
\arguments{
\item{auto}{Logical. If \code{TRUE}, includes the \code{"auto"} gene type in the
output, which has \link{plot_gene_map} determine what gene type to use
automatically.}
}
\value{
A character vector.
}
\description{
Returns a character vector containing the available gene types when plotting
\code{dna_seg} features.
}
\details{
\code{dna_segs} may contain the \code{gene_type} column, which determines the shape
of \code{dna_seg} features when they are plotted using \link{plot_gene_map}.
Elements in this column should either be one of the predefined gene types
returned by this function, or they must refer to a graphical function with
exactly the same name that returns a \code{grob} or a \code{gList} object.
}
\examples{
## To view pre-coded gene types:
gene_types()

## Load data
data(barto)
n <- length(gene_types(auto = FALSE))

## Get a small subset from the barto dataset
dna_seg <- barto$dna_segs[[3]][1:n,]
plot_gene_map(list(dna_seg))

## Change gene_types and plot again
dna_seg$gene_type <- gene_types(auto = FALSE)
dna_seg$fill <- rainbow(n)
dna_seg_r <- dna_seg
dna_seg_r$strand <- -dna_seg$strand

## Add an annotation
annot <- annotation(middle(dna_seg), text = dna_seg$gene_type, rot = 45,
                    col = dna_seg$col)

## Plot
plot_gene_map(list(dna_seg, dna_seg_r), annotations = list(annot, annot),
              annotation_height = 5, dna_seg_line = grey(0.7))

}
\seealso{
\link{plot_gene_map}, \link{dna_seg}
}
\author{
Lionel Guy
}
