% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPlotR-package.R
\docType{package}
\name{genoPlotR-package}
\alias{genoPlotR}
\alias{genoPlotR-package}
\title{genoPlotR: Plot Publication-Grade Gene and Genome Maps}
\description{
Draws gene or genome maps and comparisons between these, in a publication-grade manner. Starting from simple, common files, it will draw postscript or PDF files that can be sent as such to journals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mvpuijk.github.io/genoPlotR/}
}

}
\keyword{internal}
