% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{human_nt}
\alias{human_nt}
\title{Human-readable nucleotide scale}
\usage{
human_nt(nt, signif = FALSE)
}
\arguments{
\item{nt}{A nucleotide position.}

\item{signif}{Either a numeric or logical. If \code{FALSE}, \code{nt} is not rounded.
If this argument is numeric, it returns that amount of significant digits.}
}
\value{
Returns a list with 4 elements.
\item{n}{A numeric value corresponding to \code{nt} divided by \code{mult}
(see below).}
\item{tag}{A character, giving the multiplier used in text.}
\item{mult}{The muliplier used, in numeric value.}
\item{text}{A character, giving the value in a human readable format.}
}
\description{
Returns a human readable list from a nucleotide position or length.
}
\details{
Return a nucleotide value in nt, kb, Mb or Gb, according to the value given.
This is particularly useful to display nice scales without too many
trailing zeros.
}
\examples{
human_nt(123456)
human_nt(123456, signif = 2)
human_nt(123456890, signif = 2)

}
\author{
Lionel Guy
}
