% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPlotR.R
\docType{data}
\name{mauve_bbone}
\alias{mauve_bbone}
\title{Mauve backbone of 4 Bartonella genomes}
\format{
\code{mauve_bbone}, a list of two data frame lists, representing the
regions which are conserved in at least 2 genomes:
\describe{
\item{\code{dna_segs}}{A list of 4 \code{dna_seg} objects, containing the mauve
blocks for each genome.}
\item{\code{comparison}}{A list of 3 \code{comparison} objects.}
}
A bash script to obtain the same file as in the data is available in the
extdata folder of the package. Find its location by running
\code{system.file('extdata/mauve.sh',package = 'genoPlotR')}.

The resulting backbone file can then be read with \link{read_mauve_backbone}.
}
\usage{
data(mauve_bbone)
}
\description{
The result of a multiple genome alignment of 4 Bartonella genomes with Mauve
}
\examples{
data(mauve_bbone)
plot_gene_map(mauve_bbone$dna_segs, mauve_bbone$comparisons)

}
\references{
Mauve: https://darlinglab.org/mauve/mauve.html
}
\keyword{datasets}
