% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{middle}
\alias{middle}
\title{Find the middle point of dna_seg features}
\usage{
middle(dna_seg)
}
\arguments{
\item{dna_seg}{A \code{dna_seg} object.}
}
\value{
A numeric vector.
}
\description{
Returns a vector containing the middle point of each feature in a \code{dna_seg}
object. Useful to prepare annotations, for example.
}
\examples{
## Load data
data(barto)

## Get middles of the first dna_seg
mid <- middle(barto$dna_segs[[1]])

}
\seealso{
\link{annotation}, \link{dna_seg}
}
\author{
Lionel Guy
}
