% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{permute_dna_segs}
\alias{permute_dna_segs}
\title{Reorder dna_segs or labels to match a tree}
\usage{
permute_dna_segs(
  dna_segs,
  tree,
  exact_match = FALSE,
  return_old_labels = FALSE
)
}
\arguments{
\item{dna_segs}{Either a character vector containing \code{dna_seg} labels, or a
list of \code{dna_seg} objects.}

\item{tree}{A (phylogenetic) tree, in the form of a \code{phylo} or \code{phylog}
object, or a character string containing a file path to a Newick tree format
file.}

\item{exact_match}{Logical. If \code{TRUE}, \code{dna_seg} labels will need to match
the labels of the tree exactly. If \code{exact_match = FALSE}, tree tip labels
only need to contain the \code{dna_seg} labels for a match to be found (e.g. the
\code{dna_seg} label \code{"seq_1"} will match tree tip label \code{"E_coli_seq_1.fa"}).}

\item{return_old_labels}{Logical. If \code{TRUE}, then the \code{dna_seg} labels will
be returned using the original names provided by the \code{dna_segs} argument.
Only relevant when \code{exact_match = FALSE}, as this option can cause \code{dna_seg}
labels to be changed to match the tree tip labels.}
}
\value{
A list of \code{dna_seg} objects or a character vector of \code{dna_seg}
labels, matching the input given in the \code{dna_segs} argument.

If \code{return_old_labels = TRUE}, a list with 2 named elements will
be returned instead (\code{dna_segs}, the same return value as above,
and \code{old_labels}, a character vector of the original labels that is now
sorted).
}
\description{
Takes a list of \code{dna_seg} objects or \code{dna_seg} labels and reorganizes them
based on a given (phylogenetic) tree.
}
\details{
This function takes a character vector of \code{dna_seg} labels, either directly,
or by extracting them from a list of \code{dna_segs}, through the \code{dna_segs}
argument. Each of the labels is queried to find matching tree tip labels,
sorting them to match the order in which they are found in the tree. If
exactly 1 match is found, the \code{dna_seg} label is updated to match the tree
tip label, unless \code{exact_match = TRUE}. If multiple matches are found, an
error is returned that shows the offending \code{dna_seg} label.
}
\examples{
## Generate data
seg_labels <- c("seq_2", "seq_3", "seq_1", "seq_4")
tree_str <- paste0("(seq_1_B_bacilliformis:0.5,",
                   "(seq_2_B_grahamii:0.1,",
                   "(seq_3_B_henselae:0.1,",
                   "seq_4_B_quintana:0.2):0.1):0.1);")
tree <- ade4::newick2phylog(tree_str)

## Reorder and rename dna_seg labels to match tree
seg_labels
seg_labels <- permute_dna_segs(dna_segs = seg_labels, tree = tree)
seg_labels

}
\seealso{
\link{trim_tree}, \link{plot_gene_map}
}
\author{
Mike Puijk
}
