% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R, R/comparison.R, R/dna_seg.R
\name{range}
\alias{range}
\alias{range.annotation}
\alias{range.comparison}
\alias{range.dna_seg}
\title{Range calculation}
\usage{
\method{range}{annotation}(x, ...)

\method{range}{comparison}(x, overall = TRUE, ...)

\method{range}{dna_seg}(x, ...)
}
\arguments{
\item{x}{An object to calculate the range from.}

\item{...}{Unused.}

\item{overall}{Used only when \code{x} is a \code{comparison} object, a logical. If
\code{FALSE}, will calculate the range of each side separately instead of
calculating the overall range of \code{x}.}
}
\value{
A numeric of length 2, unless \code{x} is a \code{comparison} object
and \code{overall} is \code{FALSE}, in which case it will be a \code{data.frame} object with
two rows and two named columns, \code{xlim1} and \code{xlim2}
}
\description{
Calculate the range of \code{dna_seg}, \code{comparison} or \code{annotation} objects.
}
\examples{
## Create dna_segs and comparison
df1 <- data.frame(name = c("feat1", "feat2", "feat3"),
                  start = c(2, 1000, 1050),
                  end = c(600, 800, 1345),
                  strand = c(-1, -1, 1))
df2 <- data.frame(name = c("feat1", "feat2", "feat3"),
                  start = c(50, 800, 1200),
                  end = c(900, 1100, 1322),
                  strand = c(-1, 1, 1))
dna_seg1 <- dna_seg(df1)
dna_seg2 <- dna_seg(df2)
df3 <- data.frame(start1 = dna_seg1$start,
                  end1 = dna_seg1$end,
                  start2 = dna_seg2$start,
                  end2 = dna_seg2$end)
comparison1 <- comparison(df3)

## Range calculation
range(dna_seg1)
range(comparison1)
range(comparison1, overall = FALSE)

}
\seealso{
\link{dna_seg}, \link{comparison}, \link{annotation}, \link{trim}
}
\author{
Lionel Guy
}
