% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_other.R
\name{read_offsets}
\alias{read_offsets}
\title{Parse \code{dna_seg} offset values for plotting}
\usage{
read_offsets(offsets)
}
\arguments{
\item{offsets}{A character string with offset values to parse. See details.}
}
\value{
A list of offset values that can be passed directly to the \code{offsets}
argument of \link{plot_gene_map}. Each element of the list is a numeric vector
that corresponds to the offset values of one \code{dna_seg}.
}
\description{
This function takes a character string and separates it to parse offset
values for plotting in \link{plot_gene_map}. The offset values determine the
amount of space before each subsegment of each plotted \code{dna_seg}.
}
\details{
The format of \code{offsets} corresponds to the output of the \code{print_offsets}
argument of \link{plot_gene_map}.
\code{offsets} must be a character string containing 1 or more offset values for
each \code{dna_seg}. \code{dna_segs} are separated by commas, in plotting order,
with their subsegments separated by spaces. For each \code{dna_seg}, if a single
value is provided, then that value will become the offset for its first
subsegment. If more values are provided for a \code{dna_seg}, then there must
be as many values as there are subsegments for that dna_seg.',
}
\examples{
## Parse offset values for 2 dna_segs
## dna_seg 1: its 2 subsegments will be offset by 150 and 1300, respectively
## dna_seg 2: only its first subsegment  will be offset by 300
read_offsets("150 1300,300")

}
\seealso{
\link{read_offsets_file}, \link{plot_gene_map}
}
\author{
Mike Puijk
}
