% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_other.R
\name{read_offsets_file}
\alias{read_offsets_file}
\title{Parse \code{dna_seg} offset values for plotting from a file}
\usage{
read_offsets_file(offsets)
}
\arguments{
\item{offsets}{A character string containing a file path, or a file
connection. See details for file format.}
}
\value{
A list of offset values that can be passed directly to the \code{offsets}
argument of \link{plot_gene_map}. Each element of the list is a numeric vector
that corresponds to the offset values of one \code{dna_seg}.
}
\description{
This function parses a file containing offset
values for plotting in \link{plot_gene_map}. The offset values determine the
amount of space before each subsegment of each plotted \code{dna_seg}.
}
\details{
The file provided by \code{offsets} must contain 1 or more offset values for
each \code{dna_seg}. Each line must contain offset values for 1 \code{dna_seg}, in
plotting order, with their subsegments separated by spaces. For each
\code{dna_seg}, if a single value is provided, then that value will become the
offset for its first subsegment. If more values are provided for a \code{dna_seg},
then there must be as many values as there are subsegments for that
\code{dna_seg}.
}
\examples{
## Parse offset values for 2 dna_segs from a file
offsets_file <- system.file('extdata/offsets.txt', package = 'genoPlotR')

## When this file is parsed:
## dna_seg 1: its 2 subsegments will be offset by 150 and 1300, respectively
## dna_seg 2: only its first subsegment  will be offset by 300

## Example of an offsets file:
cat(readLines(offsets_file), sep = "\n")

## Parse offsets file
read_offsets_file(offsets_file)

## This is equivalent to:
read_offsets("150 1300,300")

}
\seealso{
\link{read_offsets}, \link{plot_gene_map}
}
\author{
Mike Puijk
}
