% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_other.R
\name{read_xlims}
\alias{read_xlims}
\title{Parse \code{xlims}, a set of \code{dna_seg} positions for plotting}
\usage{
read_xlims(xlims)
}
\arguments{
\item{xlims}{A character string with \code{xlims} to parse. See details.}
}
\value{
A list of \code{xlims} that can be passed directly to the \code{xlims}
argument of \link{plot_gene_map}. Each element of the list is a numeric vector
that corresponds to the \code{xlims} of one \code{dna_seg}.
}
\description{
This function takes a character string containing a set of \code{xlims}, positions
to print for \code{dna_segs}. Multiple start and end positions can be provided for
each \code{dna_seg}, forming subsegments to plot.
}
\details{
The format of \code{xlims} corresponds to the output of the \code{print_xlims}
argument of \link{plot_gene_map}.
\code{xlims} must be a character string containing a set of \code{xlims} for each
\code{dna_seg}. For each \code{dna_seg}, start and end positions must be provided for
each subsegment, all separated by spaces. The \code{dna_segs} themselves are
separated by commas, in plotting order. An example of the format is provided
below.
}
\examples{
## Parse xlims for 2 dna_segs
## For dna_seg 1: 2 subsegments, 1st starts at 1 and ends at 10000
## while the 2nd subsegment starts at 155000 and ends at 154000
## For dna_seg 2: 1 subsegment, starting at 500 and ending at 10500
read_xlims("1 10000 155000 154000,500 10500")

}
\seealso{
\link{read_xlims_file}, \link{plot_gene_map}
}
\author{
Mike Puijk
}
