% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_other.R
\name{read_xlims_file}
\alias{read_xlims_file}
\title{Parse \code{xlims} from a file, a set of \code{dna_seg} positions for plotting}
\usage{
read_xlims_file(dna_segs, xlims, reorder_dna_segs = FALSE)
}
\arguments{
\item{dna_segs}{Either a character vector containing \code{dna_seg} labels, or a
list of \code{dna_seg} objects.}

\item{xlims}{A character string containing a file path, or a file
connection. See details for file format.}

\item{reorder_dna_segs}{Logical. If \code{TRUE}, then the order of \code{dna_segs} will
be changed to match the order in which they first appear in the \code{xlims} file.
Additionally, any \code{dna_seg} that was not present in the \code{xlims} file will be
removed. Then, this function will return a list with the reordered
\code{dna_segs} and parsed \code{xlims} as elements, instead of just the \code{xlims}.}
}
\value{
A list of \code{xlims} that can be passed directly to the \code{xlims}
argument of \link{plot_gene_map}. Each element of the list is a numeric vector
that contains the positions to plot of one \code{dna_seg}.

If \code{reorder_dna_segs = TRUE}, a list with 2 named elements will
be returned instead (\code{xlims}
and either \code{dna_segs} or \code{seg_labels}, depending on the \code{dna_segs} argument).
}
\description{
This function parses a file containing a set of \code{xlims}, positions
to print for \code{dna_segs}. Multiple start and end positions can be provided for
each \code{dna_seg}, forming subsegments to plot.
}
\details{
The file format of \code{xlims} corresponds to the output of the \code{outfile_xlims}
argument of \link{plot_gene_map}. The format in question is a
tab-separated file containing \code{xlims} for each \code{dna_seg}. Each row
represents one subsegment and has 3 columns, containing the start position,
end position, and \code{dna_seg} label, respectively. An example of the format:

\tabular{lllll}{
x0    \tab x1    \tab seg_label\cr
2000  \tab 12000 \tab Genome_2 \cr
1000  \tab 10000 \tab Genome_1 \cr
6000  \tab 16000 \tab Genome_3 \cr
30000 \tab 40000 \tab Genome_2 \cr
}
}
\examples{
## Parse xlims for 3 dna_segs from a file
dna_seg_labels <- c("Genome_1", "Genome_2", "Genome_3")
xlims_file <- system.file('extdata/xlims.tab', package = 'genoPlotR')

## Example of an xlims file:
cat(readLines(xlims_file), sep = "\n")

## Parse xlims file, keep order provided by dna_seg_labels
read_xlims_file(dna_seg_labels, xlims_file)

## Parse xlims file, reorder dna_segs based on the xlims file
xlims_read <- read_xlims_file(dna_seg_labels, xlims_file, 
                              reorder_dna_segs = TRUE)
xlims_read$xlims
xlims_read$seg_labels

}
\seealso{
\link{read_xlims}, \link{plot_gene_map}
}
\author{
Mike Puijk
}
