% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{reverse}
\alias{reverse}
\alias{reverse.default}
\alias{reverse.dna_seg}
\alias{reverse.comparison}
\title{Reverse objects}
\usage{
reverse(x, ...)

\method{reverse}{default}(x, ...)

\method{reverse}{dna_seg}(x, ...)

\method{reverse}{comparison}(x, side = 0, ...)
}
\arguments{
\item{x}{An object to reverse.}

\item{...}{Unused.}

\item{side}{Used only when \code{x} is a \code{comparison} object, the side that
should be reversed. If \code{side = 1}, the first side will be reversed. If
\code{side = 2}, the second side will be reversed. If \code{side < 1}, no side is
reversed. If \code{side > 2}, both sides are reversed.}
}
\value{
An object with the same type as \code{x}.
}
\description{
Reverses objects, mainly meant for \code{dna_seg} and \code{comparison} objects.
}
\examples{
## load data
data(three_genes)
dna_segs <- three_genes$dna_segs
comparisons <- three_genes$comparisons

## on dna_seg
dna_segs[[1]]
reverse(dna_segs[[1]])
## on comparison
reverse(comparisons[[2]], side = 1)
reverse(comparisons[[2]], side = 3)

## With mauve backbone
data(mauve_bbone)
## Plot
plot_gene_map(dna_segs = mauve_bbone$dna_segs,
              comparisons = mauve_bbone$comparisons,
              alpha_comparisons = 0.4)

## Reverse B_bacilliformis, and the corresponding comparison (first "side")
mauve_bbone$dna_segs[[1]] <- reverse(mauve_bbone$dna_segs[[1]])
mauve_bbone$comparisons[[1]] <- reverse(mauve_bbone$comparisons[[1]], 1)
plot_gene_map(dna_segs = mauve_bbone$dna_segs,
              comparisons = mauve_bbone$comparisons,
              alpha_comparisons = 0.4)

}
\seealso{
\link{dna_seg}, \link{comparison}
}
\author{
Lionel Guy
}
