% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPlotR.R
\docType{data}
\name{three_genes}
\alias{three_genes}
\title{Three genes data set}
\format{
\code{three_genes}, a list of two data frame lists, representing three
genes in three DNA segments:
\describe{
\item{\code{dna_segs}}{A list of 3 \code{dna_seg} objects, containing three genes
(rows) each.}
\item{\code{comparisons}}{A list of 2 \code{comparison} objects.}
}
}
\usage{
data(three_genes)
}
\description{
A set of three made-up genes, compared in three chromosomes.
}
\examples{
data(three_genes)
plot_gene_map(three_genes$dna_segs, three_genes$comparisons)

}
\keyword{datasets}
