% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{trim_tree}
\alias{trim_tree}
\title{Trim a tree to remove unused sequences}
\usage{
trim_tree(dna_segs, tree, exact_match = FALSE)
}
\arguments{
\item{dna_segs}{Either a character vector containing \code{dna_seg} labels, or a
list of \code{dna_seg} objects.}

\item{tree}{A (phylogenetic) tree, in the form of a \code{phylo} or \code{phylog}
object, or a character string containing a file path to a Newick tree format
file.}

\item{exact_match}{Logical. If \code{TRUE}, \code{dna_seg} labels will need to match
the labels of the tree exactly. If \code{exact_match = FALSE}, tree tip labels
only need to contain the \code{dna_seg} labels for a match to be found (e.g. the
\code{dna_seg} label \code{"seq_1"} will match tree tip label \code{"E_coli_seq_1.fa"}).}
}
\value{
A phylog object containing the filtered tree.
}
\description{
Takes a (phylogenetic) tree and removes all tree tips that are not found
in a given set of names.
}
\details{
This function takes a character vector of \code{dna_seg} labels, either directly,
or by extracting them from a list of \code{dna_segs}, through the \code{dna_segs}
argument. Each of the labels is queried to find matching tree tip labels,
and any tree tip label without a match will removed. If multiple matches are
found for a single \code{dna_seg} label, an
error is returned that shows the offending \code{dna_seg} label.
}
\examples{
## Generate data
names <- c("seq_1", "seq_2", "seq_3")
tree_str <- paste0("(seq_1_B_bacilliformis:0.5,",
                   "(seq_2_B_grahamii:0.1,",
                   "(seq_3_B_henselae:0.1,",
                   "seq_4_B_quintana:0.2):0.1):0.1);"
                   )
tree <- ade4::newick2phylog(tree_str)

## Filter tree
tree$tre
tree <- trim_tree(dna_segs = names, tree = tree)
tree$tre

}
\seealso{
\link{permute_dna_segs}, \link{plot_gene_map}
}
\author{
Mike Puijk
}
