% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_data.R
\name{update_comparisons}
\alias{update_comparisons}
\title{Update comparisons using dna_segs}
\usage{
update_comparisons(
  dna_seg_input,
  comparison_input,
  seg_id = "locus_id",
  comparison_id = "auto",
  update_positions = FALSE,
  update_region_plot = TRUE,
  color_var = NULL,
  default_color = "grey80",
  update_from_top = TRUE
)
}
\arguments{
\item{dna_seg_input}{Either a single \code{dna_seg} or a list of \code{dna_seg}
objects.}

\item{comparison_input}{Either a single \code{comparison} or a list of
\code{comparison} objects.}

\item{seg_id}{The name of a \code{dna_seg} column, whose values will be used to
make the links to the comparisons.}

\item{comparison_id}{The shared name of the \code{comparison} columns, whose
values will be used to make the links to the \code{dna_segs}. See details.}

\item{update_positions}{Logical. If \code{TRUE}, updates the plotted positions
of the comparisons to match the \code{dna_segs}. \code{start1} and \code{end1} will be
updated using the \code{dna_seg} above the \code{comparison} in plotting order,
while \code{start2} and \code{end2} will be updated using the \code{dna_seg} under the
\code{comparison} in plotting order.}

\item{update_region_plot}{Logical. If \code{TRUE}, updates the \code{region_plot}
column of the comparisons. This only serves the purpose of being able to
transfer this information to other \code{dna_segs}, the \code{region_plot} column has
no other function in \code{comparison} objects.}

\item{color_var}{The color column to update the comparisons with. If it is
\code{col} or \code{fill}, then that column will be taken from the \code{dna_segs} to update
the \code{col} column in the comparisons. Must be either \code{col}, \code{fill}, or left as
the default \code{NULL}, which will result in no color column being updated.}

\item{default_color}{A character string providing the default color of the
\code{dna_segs}, must be either \code{NULL} or a valid color. The color given by
this argument will be ignored when updating, never overwriting any color
in the comparisons.}

\item{update_from_top}{Logical. If \code{TRUE}, updates the comparisons with
the \code{dna_segs} above them in the plotting order. Setting this to \code{FALSE}
will make it update the comparisons using the \code{dna_segs} below them in the
plotting order, but this also means that the \code{dna_segs} and comparisons
need to be provided in reverse plotting order. Therefore, it is recommended
to be used only when providing a single \code{dna_seg} and \code{comparison}.}
}
\value{
Either a single \code{comparison} object or a list of
\code{comparison} objects, matching the input given using \code{comparison_input}.
}
\description{
A \code{comparison} or list of \code{comparisons} is updated using a matching
list of \code{dna_segs}. This can be used to update the \code{region_plot}, \code{col},
or position columns of the comparisons, by linking them with \code{dna_seg}
features.
}
\details{
If \code{dna_seg_input} is a single \code{dna_seg}, it will be updated using the first
(or only) \code{comparison} from \code{comparison_input}.

The objects are linked together through shared
values. The columns for these shared values are determined by the \code{seg_id}
and \code{comparison_id} arguments, for the \code{dna_segs} and comparisons,
respectively. \code{comparison_id} refers to 2 columns, and defaults to \code{"auto"},
which will attempt to determine which columns to use automatically.
If for example, \code{comparison_id} is set as \code{"name"}, it will look for the
\code{"name1"} and \code{"name2"} columns to match to the \code{seg_id} in the \code{dna_segs}
above, and under it, respectively.

When \code{update_from_top} is \code{TRUE}, it assumes the list of \code{dna_segs} and
comparisons are in plotting order. The first \code{dna_seg} will not be updated
as there is no \code{comparison} above it to update it with. If instead
\code{update_from_top} is \code{FALSE}, the \code{dna_segs} and comparisons would have to
be supplied in reverse plot order, which is why this is not recommended.
In this case the last \code{dna_seg} in plot order will not be updated, as there
was no \code{comparison} below it to update it with.
}
\examples{
## Prepare dna_seg
names1 <- c("1A", "1B", "1C")
names2 <- c("2A", "2C", "2B")
names3 <- c("3B", "3A", "3C")

## Make dna_segs
dna_seg1 <- dna_seg(data.frame(name = names1,
                               start = (1:3) * 3,
                               end = (1:3) * 3 + 2,
                               strand = rep(1, 3)))
dna_seg2 <- dna_seg(data.frame(name = names2,
                               start = (1:3) * 3,
                               end = (1:3) * 3 + 2,
                               strand = rep(1, 3)))
dna_seg3 <- dna_seg(data.frame(name = names3,
                               start = (1:3) * 3,
                               end = (1:3) * 3 + 2,
                               strand = rep(1, 3)))
dna_segs <- list("Genome 1" = dna_seg1,
                 "Genome 2" = dna_seg2,
                 "Genome 3" = dna_seg3)

## Add colors based on the presence of a string in the feature names
id_grep <- c("A", "B")
fill_grep <- c("red", "blue")
dna_segs <- edit_dna_segs(dna_seg_input = dna_segs,
                          ids = data.frame(id = id_grep, fill = fill_grep))

## Make comparisons
comp1 <- comparison(data.frame(start1 = c(3, 6, 8), end1 = c(5, 8, 10),
                               start2 = c(3, 9, 7), end2 = c(5, 11, 9),
                               name1 = c("1A", "1B", "1C"),
                               name2 = c("2A", "2B", "2C"),
                               direction = c(1, 1, 1)))
comp2 <- comparison(data.frame(start1 = c(3, 9, 7), end1 = c(5, 11, 9),
                               start2 = c(6, 3, 8), end2 = c(8, 5, 10),
                               name1 = c("2A", "2B", "2C"),
                               name2 = c("3A", "3B", "3C"),
                               direction = c(1, 1, 1)))

## Before using update_comparisons
plot_gene_map(dna_segs = dna_segs, 
              comparisons = list(comp1, comp2),
              alpha_comparisons = 0.6)

## Apply update_comparisons to fix the positions and update the colors
comps <- update_comparisons(dna_seg_input = dna_segs,
                            comparison_input = list(comp1, comp2),
                            seg_id = "name",
                            color_var = "fill",
                            update_positions = TRUE)

## After using update_comparisons
plot_gene_map(dna_segs = dna_segs, 
              comparisons = comps,
              alpha_comparisons = 0.6)

}
\seealso{
\link{update_dna_segs}, \link{sequential_updates}, \link{dna_seg}, \link{comparison}
}
\author{
Mike Puijk
}
